refactor: extract god module (InstructorService) into sub-services (Facad pattern)
- basta
